Improvements:
1.
- Row and Column list of hits information is now only stored if there are non-zero hits on that Row or column.
*****

2.
- TTrack memeber variable fhits contains list of selected THit for that track.
- TEvent member variable fhits replace by funusedhits. This variable contain all hits that are not in any TTrack fhits list.

In this way now list of hits in Event and in Track are complementary.
deletion of THit is handled by each object either in the TTrack or in the TEvent destructors.

AnalysisBase::DrawSelection modified to take account this differences.
***

3.
Condition to not overwrite files modified. Now the code runs if the test mode is on, and it does not create any new empty output file in case it doesn't already exist.

Implemented overwrite flag to explicitly overwrite file if it exist. If it doesn't exist continue execution normally creating a new file.
***